move tag files to buildtag package
